requirements.txt
blinker==1.7.0
click==8.1.7
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
scikit-learn==1.2.2
scipy==1.12.0
six==1.16.0
threadpoolctl==3.2.0
tzdata==2024.1
Werkzeug==3.0.1
